+++ backend/src/app.module.ts
@@ -3,12 +3,12 @@ import { TypeOrmModule, TypeOrmModuleOptions } from "@nestjs/typeorm";
 
 import { ValidatorService } from "./services/validator.service";
 import { BotsService } from "./services/bots.service";
-import { HighscoresController } from "./controllers/highscore/highscores.controller";
-import { BoardsController } from "./controllers/boards/boards.controller";
-import { BotsController } from "./controllers/bots/bots.controller";
-import { SeasonsController } from "./controllers/seasons/seasons.controller";
-import { SlackController } from "./controllers/slack/slack.controller";
-import { TeamsController } from "./controllers/teams/teams.controller";
+import { HighscoresController } from "./controllers/highscores.controller";
+import { BoardsController } from "./controllers/boards.controller";
+import { BotsController } from "./controllers/bots.controller";
+import { SeasonsController } from "./controllers/seasons.controller";
+import { SlackController } from "./controllers/slack.controller";
+import { TeamsController } from "./controllers/teams.controller";
 import { CustomLogger } from "./logger";
 import { BoardsService } from "./services/board.service";
 import { HighScoresService } from "./services/high-scores.service";
@@ -25,7 +25,7 @@ import { SlackService } from "./services/slack.service";
 import { TeamsService } from "./services/teams.service";
 import { RecordingsService } from "./services/recordings.service";
 import { RecordingsEntity } from "./db/models/recordings.entity";
-import { RecordingsController } from "./controllers/recordings/recordings.controller";
+import { RecordingsController } from "./controllers/recordings.controller";
 import { RecordingsRepository } from "./db/repositories/recordings.repository";
 import { BoardConfigService } from "./services/board-config.service";
 import { BoardConfigEntity } from "./db/models/boardConfig.entity";
