+++ backend/src/controllers/bots.controller.ts
