+++ backend/src/controllers/highscores.controller.ts
