+++ backend/src/controllers/recordings.controller.ts
@@ -1,8 +1,8 @@
 import { Controller, Get, Body, Post, Param } from "@nestjs/common";
 import { ApiUseTags, ApiResponse } from "@nestjs/swagger";
-import { RecordingsService } from "../../services/recordings.service";
-import { RecordingListDto } from "../../models/recording-list.dto";
-import { RecordingPublicDto } from "../../models/recording-public.dto";
+import { RecordingsService } from "../services/recordings.service";
+import { RecordingListDto } from "../models/recording-list.dto";
+import { RecordingPublicDto } from "../models/recording-public.dto";
 
 @ApiUseTags("Recordings")
 @Controller("api/recordings")
